requirements.txt
transformers>=4.39.0
torch>=2.0.0
sentence-transformers>=2.2.2

faiss-cpu>=1.7.4

langchain-community
langchain-core
langchain-huggingface


pypdf>=3.17.4
markdown>=3.5.2

streamlit>=1.35.0
fastapi>=0.110.0
uvicorn>=0.29.0

python-dotenv>=1.0.1
tqdm>=4.66.4

rich>=13.7.1